Update README.md
Total points: 13.

For this assignment we will be based on the following database hr:

*Provided by w3resource website

It’s good practice to make SQL keywords uppercase to distinguish them from other parts of your query. Please do this to make it easier for the grader. Also remember to include a semicolon at the end of your query. This tells SQL where the end of your query is!

    [1 pt] Provide the SQL code to display the names (first_name, last_name separately) using alias names FirstName and LastName, respectively, from the employees table.

    [1 pt] Provide the SQL code to display the unique department ID from the employees table.

    [1.5 pts] Provide the SQL code to display the first and last names, salary and PF of all the employees (PF is calculated as 15% of salary) of all employees according to their first name in descending order.

    [1.5 pts] Provide the SQL code to display the name (first_name, last_name) and salary for all employees whose salary is not in the range $10,000 through $15,000 and are in department 30 or 100.